perf(q4): fp16 block scale — q4_0 blocks 20 -> 18 bytes (-10% int4 weight traffic)
Perf item 2 (plan_v4 E2.ter): the q4_0 per-block scale was f32; store it as IEEE
fp16 bits instead (llama.cpp q4_0 layout). Storage-only — all math stays f32,
one convert per 32-weight block (native __fp16 on aarch64, bit-exact portable
fallback elsewhere, subnormals handled). The quantizer rounds nibbles against
the ROUNDTRIPPED scale so dequant reads exactly the stored value.

Field renamed scale -> scale_f16 so every reader was compiler-caught: all C
kernels (scalar/NEON/SDOT/AVX2/AVX-512-VNNI matvec, matmat, argmax), the Metal
q4blk twin (MSL half), and the CUDA q4blk twin (__half; compile-checked via
make check-isa for x86/ARM — nvcc build untested until a GPU box).

Measured (M1):
- quality: teacher-forced int4 ladder 43.79% -> 42.99% vs bf16 rails (-0.8pp,
  negligible; int8/bf16 paths untouched, test-golden ALL PASS)
- perf: 0.6B --int4 -j4 interleaved A/B: Talker ~-3%, CP ~-4.5%, RTF to 0.51
- --self-test PASS all shapes; Metal gpu-selftest matvec_q4_0 PASS; check-isa OK
- bonus: -10% RAM for q4 weights (18/20 B per block)

Also: fix test-emotion-ft silent SKIP — EXPR_FT pointed at the stale
presets/expr/italian_csp.expr name; now italian_csp_topk6.expr (the shipped WIN
pack), so the emotion-FT smoke actually runs in test-all again (caught while
auditing a suspiciously fast test-all).

+++ qwen_tts_cuda_talker.cu
@@ -23,6 +23,7 @@
 
 #include <cuda_runtime.h>
 #include <cuda_bf16.h>
+#include <cuda_fp16.h>   /* __half q4blk scale (matches q4_0_block_t fp16 layout) */
 #include <stdio.h>
 #include <stdlib.h>
 #include <math.h>
@@ -54,9 +55,9 @@ __global__ void k_matvec_bf16(const __nv_bfloat16 *W, const float *x, float *y, 
     if (lane == 0) y[row] = s;
 }
 
-/* q4_0 block: 32 weights = fp32 scale + 16 bytes (low nibble=even idx, high=odd), val=(nib-8)*scale.
- * MUST match qwen_tts_kernels.h q4_0_block_t exactly. */
-typedef struct { float scale; unsigned char qs[16]; } q4blk;
+/* q4_0 block: 32 weights = fp16 scale + 16 bytes (low nibble=even idx, high=odd), val=(nib-8)*scale.
+ * MUST match qwen_tts_kernels.h q4_0_block_t exactly (fp16 scale, 18 B/block since perf item 2). */
+typedef struct { __half scale; unsigned char qs[16]; } q4blk;
 /* warp-per-row q4_0 matvec: int4 weight (0.5 byte) × f32 activation. Half the bytes of int8. */
 __global__ void k_matvec_q4_0(const q4blk *W, const float *x, float *y, int rows, int cols) {
     int row = (blockIdx.x*blockDim.x + threadIdx.x) >> 5;
@@ -69,7 +70,7 @@ __global__ void k_matvec_q4_0(const q4blk *W, const float *x, float *y, int rows
         const q4blk *b = wr + (c>>5); int ic = c & 31;
         unsigned char byte = b->qs[ic>>1];
         int nib = (ic&1) ? (byte>>4) : (byte&0x0F);
-        s += (float)(nib-8) * b->scale * x[c];
+        s += (float)(nib-8) * __half2float(b->scale) * x[c];
     }
     #pragma unroll
     for (int o = 16; o > 0; o >>= 1) s += __shfl_down_sync(0xffffffffu, s, o);
@@ -380,7 +381,7 @@ __global__ void k_matmat_q4_0(const q4blk *W,const float *X,float *Y,int rows,in
     for(int b=0;b<QB_MAX;++b) s[b]=0.f;
     for(int c=lane;c<cols;c+=32){ const q4blk *bk=wr+(c>>5); int ic=c&31;
         unsigned char byte=bk->qs[ic>>1]; int nib=(ic&1)?(byte>>4):(byte&0x0F);
-        float w=(float)(nib-8)*bk->scale;
+        float w=(float)(nib-8)*__half2float(bk->scale);
         for(int b=0;b<B;++b) s[b]+=w*X[(size_t)b*cols+c]; }
     for(int b=0;b<B;++b){ float v=s[b];
         #pragma unroll
